Move UserTextInformationFrame to its own file
